+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## **2.0.1** - *2020-04-01*
+* ability to use MongoDB as session storage added;
+* consider `FACADE_PORT`, `DASHBOARD_PORT`, `API_PORT` env vars;
+* `Facade`:
+    * retry logic added for failed api requests;
+    * api.js additional checks added;
+* `Dashboard/Gallery` - handle xmp-js errors
+
+
 ## **2.0.0** - *2020-03-29*
 Break the rock - split the monolith into smaller parts: `api`, `dashboard` and `facade`.